+++ api/CHANGELOG.md
@@ -14,6 +14,8 @@
 - Added a button to the installation in progress dialog on the patcher screen allowing the user to "Leave anyway" and not wait for the installer to finish or timeout/fail
 - Added a "External installer", "Rooted mount installer", "System installer" and "Shizuku" installation types to the app info page for saved patched apps in the "Apps" tab
 - Added a confirmation dialog when tapping the back button during a install on the "App info" page for saved patched apps in the "Apps" tab
+- Removed the "Default" app selection page filter and replaced it with a "Installed only" and "Patches available" filter, along with making them multiselectable
+- Added new options under "Ptach selection buttons order" in Settings > Advanced which allows the user to hide patch selection page action buttons
 
 
 # Bug fixes